Move SecretData objects to config package so can be used across the app
