+++ config/src/main/java/com/quorum/tessera/config/vault/data/GetSecretData.java
@@ -1,0 +1,7 @@
+package com.quorum.tessera.config.vault.data;
+
+import com.quorum.tessera.config.KeyVaultType;
+
+public interface GetSecretData {
+    KeyVaultType getType();
+}
