+++ config/src/main/java/com/quorum/tessera/config/vault/data/SetSecretData.java
@@ -1,0 +1,7 @@
+package com.quorum.tessera.config.vault.data;
+
+import com.quorum.tessera.config.KeyVaultType;
+
+public interface SetSecretData {
+    KeyVaultType getType();
+}
