+++ key-pair-converter/src/main/java/com/quorum/tessera/keypairconverter/KeyPairConverter.java
@@ -7,15 +7,18 @@ import com.quorum.tessera.config.keypairs.AzureVaultKeyPair;
 import com.quorum.tessera.config.keypairs.ConfigKeyPair;
 import com.quorum.tessera.config.keypairs.HashicorpVaultKeyPair;
 import com.quorum.tessera.config.util.EnvironmentVariableProvider;
+import com.quorum.tessera.config.vault.data.AzureGetSecretData;
+import com.quorum.tessera.config.vault.data.GetSecretData;
+import com.quorum.tessera.config.vault.data.HashicorpGetSecretData;
 import com.quorum.tessera.encryption.KeyPair;
 import com.quorum.tessera.encryption.PrivateKey;
 import com.quorum.tessera.encryption.PublicKey;
-import com.quorum.tessera.key.vault.GetSecretDataFactory;
 import com.quorum.tessera.key.vault.KeyVaultClientFactory;
 import com.quorum.tessera.key.vault.KeyVaultService;
 import com.quorum.tessera.key.vault.KeyVaultServiceFactory;
 
-import java.util.*;
+import java.util.Base64;
+import java.util.Collection;
 import java.util.stream.Collectors;
 
 
@@ -45,36 +48,42 @@ public class KeyPairConverter {
 
             KeyVaultServiceFactory keyVaultServiceFactory = KeyVaultServiceFactory.getInstance(KeyVaultType.AZURE);
             KeyVaultClientFactory keyVaultClientFactory = KeyVaultClientFactory.getInstance(KeyVaultType.AZURE);
-            GetSecretDataFactory getSecretDataFactory = GetSecretDataFactory.getInstance(KeyVaultType.AZURE);
+//            GetSecretDataFactory getSecretDataFactory = GetSecretDataFactory.getInstance(KeyVaultType.AZURE);
 
             KeyVaultService keyVaultService = keyVaultServiceFactory.create(config, envProvider, keyVaultClientFactory);
 
             AzureVaultKeyPair akp = (AzureVaultKeyPair) configKeyPair;
-            Map<String, String> publicKeyData = Collections.singletonMap("secretName", akp.getPublicKeyId());
-            Map<String, String> privateKeyData = Collections.singletonMap("secretName", akp.getPrivateKeyId());
+//            Map<String, String> publicKeyData = Collections.singletonMap("secretName", akp.getPublicKeyId());
+//            Map<String, String> privateKeyData = Collections.singletonMap("secretName", akp.getPrivateKeyId());
 
-            base64PublicKey = keyVaultService.getSecret(getSecretDataFactory.create(publicKeyData));
-            base64PrivateKey = keyVaultService.getSecret(getSecretDataFactory.create(privateKeyData));
+            GetSecretData getPublicKeyData = new AzureGetSecretData(akp.getPublicKeyId());
+            GetSecretData getPrivateKeyData = new AzureGetSecretData(akp.getPrivateKeyId());
+
+            base64PublicKey = keyVaultService.getSecret(getPublicKeyData);
+            base64PrivateKey = keyVaultService.getSecret(getPrivateKeyData);
         }
         else if(configKeyPair instanceof HashicorpVaultKeyPair) {
 
             KeyVaultServiceFactory keyVaultServiceFactory = KeyVaultServiceFactory.getInstance(KeyVaultType.HASHICORP);
             KeyVaultClientFactory keyVaultClientFactory = KeyVaultClientFactory.getInstance(KeyVaultType.HASHICORP);
-            GetSecretDataFactory getSecretDataFactory = GetSecretDataFactory.getInstance(KeyVaultType.HASHICORP);
+//            GetSecretDataFactory getSecretDataFactory = GetSecretDataFactory.getInstance(KeyVaultType.HASHICORP);
 
             KeyVaultService keyVaultService = keyVaultServiceFactory.create(config, envProvider, keyVaultClientFactory);
 
             HashicorpVaultKeyPair hkp = (HashicorpVaultKeyPair) configKeyPair;
-            Map<String, String> publicKeyData = new HashMap<>();
-            publicKeyData.put("secretName", hkp.getPublicKeyId());
-            publicKeyData.put("secretPath", hkp.getSecretPath());
+//            Map<String, String> publicKeyData = new HashMap<>();
+//            publicKeyData.put("secretName", hkp.getPublicKeyId());
+//            publicKeyData.put("secretPath", hkp.getSecretPath());
 
-            Map<String, String> privateKeyData = new HashMap<>();
-            privateKeyData.put("secretName", hkp.getPrivateKeyId());
-            privateKeyData.put("secretPath", hkp.getSecretPath());
+//            Map<String, String> privateKeyData = new HashMap<>();
+//            privateKeyData.put("secretName", hkp.getPrivateKeyId());
+//            privateKeyData.put("secretPath", hkp.getSecretPath());
 
-            base64PublicKey = keyVaultService.getSecret(getSecretDataFactory.create(publicKeyData));
-            base64PrivateKey = keyVaultService.getSecret(getSecretDataFactory.create(privateKeyData));
+            GetSecretData getPublicKeyData = new HashicorpGetSecretData(hkp.getSecretPath(), hkp.getPublicKeyId());
+            GetSecretData getPrivateKeyData = new HashicorpGetSecretData(hkp.getSecretPath(), hkp.getPrivateKeyId());
+
+            base64PublicKey = keyVaultService.getSecret(getPublicKeyData);
+            base64PrivateKey = keyVaultService.getSecret(getPrivateKeyData);
 
         }
         else {
