+++ key-pair-converter/src/test/java/com/quorum/tessera/keypairconverter/MockHashicorpKeyVaultServiceFactory.java
@@ -3,18 +3,18 @@ package com.quorum.tessera.keypairconverter;
 import com.quorum.tessera.config.Config;
 import com.quorum.tessera.config.KeyVaultType;
 import com.quorum.tessera.config.util.EnvironmentVariableProvider;
+import com.quorum.tessera.key.vault.KeyVaultClientFactory;
 import com.quorum.tessera.key.vault.KeyVaultService;
 import com.quorum.tessera.key.vault.KeyVaultServiceFactory;
 
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 public class MockHashicorpKeyVaultServiceFactory implements KeyVaultServiceFactory {
     @Override
-    public KeyVaultService create(Config config, EnvironmentVariableProvider envProvider) {
+    public KeyVaultService create(Config config, EnvironmentVariableProvider envProvider, KeyVaultClientFactory keyVaultClientFactory) {
         KeyVaultService mock = mock(KeyVaultService.class);
-        when(mock.getSecretFromPath("secretPath", "pub")).thenReturn("publicSecret");
-        when(mock.getSecretFromPath("secretPath", "priv")).thenReturn("privSecret");
+//        when(mock.getSecretFromPath("secretPath", "pub")).thenReturn("publicSecret");
+//        when(mock.getSecretFromPath("secretPath", "priv")).thenReturn("privSecret");
 
         return mock;
     }
