+++ key-vault/azure-key-vault/src/main/java/com/quorum/tessera/key/vault/azure/AzureGetSecretData.java
@@ -1,22 +1,22 @@
-package com.quorum.tessera.key.vault.azure;
-
-import com.quorum.tessera.config.KeyVaultType;
-import com.quorum.tessera.key.vault.GetSecretData;
-
-public class AzureGetSecretData implements GetSecretData {
-
-    private String secretName;
-
-    public AzureGetSecretData(String secretName) {
-        this.secretName = secretName;
-    }
-
-    @Override
-    public KeyVaultType getType() {
-        return KeyVaultType.AZURE;
-    }
-
-    public String getSecretName() {
-        return secretName;
-    }
-}
+//package com.quorum.tessera.key.vault.azure;
+//
+//import com.quorum.tessera.config.KeyVaultType;
+//import com.quorum.tessera.key.vault.GetSecretData;
+//
+//public class AzureGetSecretData implements GetSecretData {
+//
+//    private String secretName;
+//
+//    public AzureGetSecretData(String secretName) {
+//        this.secretName = secretName;
+//    }
+//
+//    @Override
+//    public KeyVaultType getType() {
+//        return KeyVaultType.AZURE;
+//    }
+//
+//    public String getSecretName() {
+//        return secretName;
+//    }
+//}
