+++ key-vault/azure-key-vault/src/main/java/com/quorum/tessera/key/vault/azure/AzureSetSecretData.java
@@ -1,28 +1,28 @@
-package com.quorum.tessera.key.vault.azure;
-
-import com.quorum.tessera.config.KeyVaultType;
-import com.quorum.tessera.key.vault.SetSecretData;
-
-public class AzureSetSecretData implements SetSecretData {
-    private String secretName;
-
-    private String secret;
-
-    public AzureSetSecretData(String secretName, String secret) {
-        this.secretName = secretName;
-        this.secret = secret;
-    }
-
-    public String getSecretName() {
-        return secretName;
-    }
-
-    public String getSecret() {
-        return secret;
-    }
-
-    @Override
-    public KeyVaultType getType() {
-        return KeyVaultType.AZURE;
-    }
-}
+//package com.quorum.tessera.key.vault.azure;
+//
+//import com.quorum.tessera.config.KeyVaultType;
+//import com.quorum.tessera.key.vault.SetSecretData;
+//
+//public class AzureSetSecretData implements SetSecretData {
+//    private String secretName;
+//
+//    private String secret;
+//
+//    public AzureSetSecretData(String secretName, String secret) {
+//        this.secretName = secretName;
+//        this.secret = secret;
+//    }
+//
+//    public String getSecretName() {
+//        return secretName;
+//    }
+//
+//    public String getSecret() {
+//        return secret;
+//    }
+//
+//    @Override
+//    public KeyVaultType getType() {
+//        return KeyVaultType.AZURE;
+//    }
+//}
