+++ key-vault/hashicorp-key-vault/src/test/java/com/quorum/tessera/key/vault/hashicorp/HashicorpKeyVaultServiceTest.java
@@ -5,9 +5,11 @@ import com.bettercloud.vault.VaultException;
 import com.bettercloud.vault.api.Logical;
 import com.bettercloud.vault.response.LogicalResponse;
 import com.quorum.tessera.config.HashicorpKeyVaultConfig;
-import com.quorum.tessera.key.vault.GetSecretData;
+import com.quorum.tessera.config.vault.data.GetSecretData;
+import com.quorum.tessera.config.vault.data.HashicorpGetSecretData;
+import com.quorum.tessera.config.vault.data.HashicorpSetSecretData;
+import com.quorum.tessera.config.vault.data.SetSecretData;
 import com.quorum.tessera.key.vault.KeyVaultException;
-import com.quorum.tessera.key.vault.SetSecretData;
 import org.junit.Before;
 import org.junit.Test;
