+++ key-vault/hashicorp-key-vault/src/test/java/com/quorum/tessera/key/vault/hashicorp/HashicorpSetSecretDataTest.java
@@ -1,6 +1,7 @@
 package com.quorum.tessera.key.vault.hashicorp;
 
 import com.quorum.tessera.config.KeyVaultType;
+import com.quorum.tessera.config.vault.data.HashicorpSetSecretData;
 import org.junit.Before;
 import org.junit.Test;
