+++ key-vault/key-vault-api/src/main/java/com/quorum/tessera/key/vault/GetSecretData.java
@@ -1,7 +1,7 @@
-package com.quorum.tessera.key.vault;
-
-import com.quorum.tessera.config.KeyVaultType;
-
-public interface GetSecretData {
-    KeyVaultType getType();
-}
+//package com.quorum.tessera.key.vault;
+//
+//import com.quorum.tessera.config.KeyVaultType;
+//
+//public interface GetSecretData {
+//    KeyVaultType getType();
+//}
