+++ key-vault/key-vault-api/src/main/java/com/quorum/tessera/key/vault/KeyVaultService.java
@@ -1,5 +1,8 @@
 package com.quorum.tessera.key.vault;
 
+import com.quorum.tessera.config.vault.data.GetSecretData;
+import com.quorum.tessera.config.vault.data.SetSecretData;
+
 public interface KeyVaultService {
 //    String getSecret(String secretName);
 //
