+++ key-vault/key-vault-api/src/main/java/com/quorum/tessera/key/vault/SetSecretData.java
@@ -1,7 +1,7 @@
-package com.quorum.tessera.key.vault;
-
-import com.quorum.tessera.config.KeyVaultType;
-
-public interface SetSecretData {
-    KeyVaultType getType();
-}
+//package com.quorum.tessera.key.vault;
+//
+//import com.quorum.tessera.config.KeyVaultType;
+//
+//public interface SetSecretData {
+//    KeyVaultType getType();
+//}
